agriculture-ai-main/ml-service/requirements.txt
# ── Web framework ────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1

# ── Data / ML ─────────────────────────────────────
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2

# ── AI / LLM ──────────────────────────────────────
openai==1.30.1
httpx==0.27.0

# ── Vector search (local, no external DB) ─────────
faiss-cpu==1.8.0

# ── Sentence embeddings (offline fallback) ────────
sentence-transformers==3.0.1

# ── Utilities ─────────────────────────────────────
pydantic==2.7.1
python-multipart==0.0.9

# ── Legacy Flask kept for backwards compat if needed ─
flask==3.0.0
flask-cors==4.0.0
